pyproject.toml
[project]
name = "movidius"
version = "0.1.0"
description = "A simple repo for saving models onto the movidius compute stick"
requires-python = ">=3.12, <3.13"


dependencies = [
    "torch>=2.7.0",
    "onnx",
    "onnxscript",
    "onnx-simplifier",
    "openvino>=2024.0.0",
    "depthai",
    "blobconverter",
    "pytest",
    "chalk-diagrams",
    "ipython",
    "jupyter",
    "pyusb>=1.3.1",
    "onnxruntime>=1.24.2",
    "opencv-python>=4.13.0.92",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]

[tool.uv.sources]
chalk-diagrams = { git = "https://github.com/chalk-diagrams/chalk" }
# Optional configuration for uv
